Began eliminating excess vertical whitespace
Just a suggestion.  This is obviously discretionary, (and I didn't do the whole job so if you accept you should finish it) but the examples are not presenting well in discourse (some get a vertical scrollbar, which makes them hard to read there), and in a few places my internal pattern matcher is making the examples hard to parse.  Don't ask me why, but it took me hours to understand this:

```swift
func reportProgress(
  @Percentage of progress: Int
) { ... }
```

The middle line read to me as some kind of property declaration with a newfangled syntax!